@@ -11,19 +11,20 @@
   "private": true,
   "dependencies": {
     "@abacritt/angularx-social-login": "^1.2.5",
-    "@angular/animations": "~13.0.0",
+    "@angular/animations": "^14.2.9",
     "@angular/cdk": "^13.3.9",
-    "@angular/common": "~13.0.0",
-    "@angular/compiler": "~13.0.0",
-    "@angular/core": "~13.0.0",
-    "@angular/forms": "~13.0.0",
+    "@angular/common": "^14.2.9",
+    "@angular/compiler": "^14.2.9",
+    "@angular/core": "^14.2.9",
+    "@angular/forms": "^14.2.9",
     "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "~13.0.0",
-    "@angular/platform-browser-dynamic": "~13.0.0",
-    "@angular/router": "~13.0.0",
+    "@angular/platform-browser": "^14.2.9",
+    "@angular/platform-browser-dynamic": "^14.2.9",
+    "@angular/router": "^14.2.9",
     "@fortawesome/fontawesome-free": "^6.1.1",
     "@popperjs/core": "^2.11.5",
     "bootstrap": "^5.1.3",
+    "i18n": "^0.15.1",
     "jquery": "^3.6.0",
     "rxjs": "~7.4.0",
     "sweetalert2": "^11.6.5",
@@ -31,9 +32,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.9",
-    "@angular/cli": "~13.0.2",
-    "@angular/compiler-cli": "~13.0.0",
+    "@angular-devkit/build-angular": "^14.2.8",
+    "@angular/cli": "^14.2.8",
+    "@angular/compiler-cli": "^14.2.9",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.20.55",
     "jasmine-core": "~3.10.0",
@@ -42,6 +43,6 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.4.3"
+    "typescript": "~4.8.4"
   }
 }
